requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0

# Cheminformatics (SMILES processing)
# Note: RDKit is best installed via conda: conda install -c conda-forge rdkit
# If using pip, you may need: pip install rdkit-pypi
rdkit>=2023.3.1

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration Files
pyyaml>=6.0

# CLI and Output Formatting
art>=6.0
colorama>=0.4.6

# Optional: For Jupyter notebook support
# jupyter>=1.0.0
# ipython>=8.0.0

# Optional: For advanced ML models
# Uncomment to use XGBoost or LightGBM
# xgboost>=2.0.0
# lightgbm>=4.0.0

# Optional: For deep learning (future support)
# tensorflow>=2.13.0
# torch>=2.0.0

